requirements.txt
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
docstring_parser==0.16
faiss-cpu==1.8.0.post1
fonttools==4.53.1
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-aiplatform==1.63.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.64.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.0
grpcio-status==1.66.0
idna==3.8
importlib_resources==6.4.4
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.2
mplcursors==0.5.3
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
scikit-learn==1.5.1
scipy==1.13.1
shapely==2.0.6
six==1.16.0
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zipp==3.20.1
